Static variable using constructor-class
In Java, a static variable is a class-level variable, also known as a class variable. It belongs to the class itself rather than any specific instance of the class. This means that regardless of how many instances of the class are created, there will only be one copy of the static variable shared among all instances.